Do not delete content of the directory which you cannot list
Addressing:
2013-07-30 11:29:25,250 [TP-Processor1] ERROR com.redhat.rhn.frontend.servlets.SessionFilter - Error during transaction. Rolling back
javax.servlet.ServletException: java.lang.NullPointerException
	at org.apache.struts.action.RequestProcessor.processException(RequestProcessor.java:520)
	at org.apache.struts.action.RequestProcessor.processActionPerform(RequestProcessor.java:427)
	at org.apache.struts.action.RequestProcessor.process(RequestProcessor.java:228)
	at com.redhat.rhn.frontend.struts.RhnRequestProcessor.process(RhnRequestProcessor.java:102)
	at org.apache.struts.action.ActionServlet.process(ActionServlet.java:1913)
	...
Caused by: java.lang.NullPointerException
	at com.redhat.rhn.manager.audit.scap.file.ScapFileManager.deleteFilesForTestResult(ScapFileManager.java:64)
	at com.redhat.rhn.manager.audit.ScapManager.deleteScan(ScapManager.java:310)
	at com.redhat.rhn.manager.audit.ScapManager.deleteScansInSet(ScapManager.java:295)
	at com.redhat.rhn.frontend.action.systems.audit.XccdfDeleteConfirmAction.execute(XccdfDeleteConfirmAction.java:86)
	at org.apache.struts.action.RequestProcessor.processActionPerform(RequestProcessor.java:425)
	... 40 more

+++ java/code/src/com/redhat/rhn/manager/audit/scap/file/ScapFileManager.java
@@ -61,8 +61,11 @@ public class ScapFileManager {
     public static void deleteFilesForTestResult(XccdfTestResult tr) {
         File folder = new File(getStoragePath(tr));
         if (folder.exists() && folder.isDirectory()) {
-            for (File file : folder.listFiles()) {
-               file.delete();
+            File[] files = folder.listFiles();
+            if (files != null) {
+                for (File file : files) {
+                    file.delete();
+                }
             }
             folder.delete();
         }
